requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask
gunicorn
tensorflow==2.16.1
flask_cors
numpy
Pillow
keras==3.3.3
numpy
